+++ marathon-to-kubernetes-core/src/main/java/fr/insee/innovation/marathontokubernetes/core/MarathonToKubernetesApplication.java
@@ -1,0 +1,15 @@
+package fr.insee.innovation.marathontokubernetes.core;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class MarathonToKubernetesApplication{
+
+	public static void main(String[] args) {
+		SpringApplication.run(MarathonToKubernetesApplication.class, args);
+	}
+
+
+
+}
