+++ marathon-to-kubernetes-core/src/main/resources/application.properties
