+++ marathon-to-kubernetes-core/src/test/java/fr/insee/innovation/marathontokubernetes/core/MarathonToKubernetesApplicationTests.java
@@ -1,0 +1,13 @@
+package fr.insee.innovation.marathontokubernetes.core;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class MarathonToKubernetesApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
