+++ marathon-to-kubernetes-core/src/test/resources/marathon/adventofcodeleaderboard.json
