+++ marathon-to-kubernetes-core/src/test/resources/marathon/drawio.json
