+++ marathon-to-kubernetes-core/src/test/resources/marathon/privileged.json
