+++ marathon-to-kubernetes-java-cli/src/main/java/fr/insee/innovation/marathontokubernetesjavacli/MarathonToKubernetesJavaCliApplication.java
@@ -1,0 +1,13 @@
+package fr.insee.innovation.marathontokubernetesjavacli;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class MarathonToKubernetesJavaCliApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(MarathonToKubernetesJavaCliApplication.class, args);
+	}
+
+}
