+++ marathon-to-kubernetes-java-cli/src/main/resources/application.properties
@@ -1,0 +1,1 @@
+
