+++ pom.xml
@@ -1,16 +1,23 @@
 <?xml version="1.0" encoding="UTF-8" ?>
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 https://maven.apache.org/xsd/maven-4.0.0.xsd">
+<project xmlns="http://maven.apache.org/POM/4.0.0" 
+	xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 https://maven.apache.org/xsd/maven-4.0.0.xsd">
 	<modelVersion>4.0.0</modelVersion>
+	
+	
 	<parent>
 		<groupId>org.springframework.boot</groupId>
 		<artifactId>spring-boot-starter-parent</artifactId>
 		<version>2.2.2.RELEASE</version>
-		<relativePath /> <!-- lookup parent from repository -->
+		<relativePath />
 	</parent>
+
+
 	<groupId>fr.insee.innovation</groupId>
-	<artifactId>marathon-to-kubernetes</artifactId>
+	<artifactId>marathon-to-kubernetes-parent</artifactId>
 	<version>0.0.1-SNAPSHOT</version>
-	<name>marathon-to-kubernetes</name>
+	<name>marathon-to-kubernetes-parent</name>
+	<packaging>pom</packaging>
+
 	<description>Convert marathon.json to kubernetes yml</description>
 
 	<properties>
@@ -34,20 +41,6 @@
 				</exclusion>
 			</exclusions>
 		</dependency>
-
-		<dependency>
-			<groupId>io.fabric8</groupId>
-			<artifactId>kubernetes-client</artifactId>
-			<version>4.6.4</version>
-		</dependency>
-
-		<!-- https://mvnrepository.com/artifact/com.mesosphere/marathon-client -->
-		<dependency>
-			<groupId>com.mesosphere</groupId>
-			<artifactId>marathon-client</artifactId>
-			<version>0.6.2</version>
-		</dependency>
-
 	</dependencies>
 
 	<build>
@@ -76,5 +69,10 @@
 			</plugin>
 		</plugins>
 	</build>
+
+	<modules>
+		<module>marathon-to-kubernetes-core</module>
+		<module>marathon-to-kubernetes-java-cli</module>
+	</modules>
 
 </project>
