--- src/main/java/fr/insee/innovation/marathontokubernetes/configuration/CustomObjectMapper.java
@@ -1,18 +1,0 @@
-package fr.insee.innovation.marathontokubernetes.configuration;
-
-import com.fasterxml.jackson.databind.DeserializationFeature;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class CustomObjectMapper {
-
-
-    @Bean
-    public ObjectMapper getMapper() {
-        ObjectMapper objectMapper = new ObjectMapper();
-        objectMapper.configure(DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES, false);
-        return objectMapper;
-    }
-}
