--- src/test/java/fr/insee/innovation/marathontokubernetes/MarathonToKubernetesApplicationTests.java
@@ -1,13 +1,0 @@
-package fr.insee.innovation.marathontokubernetes;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class MarathonToKubernetesApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
